Extrait l'aggregateur_metrique de la partie
